+++ tests/FsSpreadsheet.Js.Tests/DefaultIO.Tests.fs
@@ -2,8 +2,9 @@
 
 open TestingUtils
 open FsSpreadsheet
-open FsSpreadsheet.Exceljs
+open FsSpreadsheet.Js
 open Fable.Core
+
 
 let private readFromTestFile (testFile: DefaultTestObject.TestFiles) =
     FsWorkbook.fromXlsxFile(testFile.asRelativePathNode)